Remove CI parallel limit to save user waiting time
I am not sure whether we add this limitation before, but
I see our CI will finish after 20 min less or more. it
because we have 5 python matrix and only 4 parallel slots.

It is open source project with no limitation for the github
action run time, so maybe we could directly remove the parallel
for ci